requirements.txt
beautifulsoup4
pypdf
gradio
langchain[llms]
python-dotenv
openai
tiktoken